requirements.txt
accelerate==0.20.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.0.1
bitsandbytes==0.42.0
blinker==1.9.0
brotli==1.2.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
coloredlogs==15.0.1
cryptography==46.0.3
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==8.1.1
durationpy==0.9
ecdsa==0.19.1
email_validator==2.2.0
faiss-cpu==1.7.4
fastapi==0.115.9
fastapi-cli==0.0.7
ffmpy==1.0.0
filelock==3.18.0
firecrawl-py==1.15.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==3.0.1
flatbuffers==25.2.10
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
gpt4all==2.8.2
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.4.2
jose==1.0.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.4.0
kubernetes==32.0.1
langchain>=0.3.0
langchain-chroma>=0.1.4
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-nomic>=0.1.4
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
langchain-text-splitters>=0.3.0
langchainhub>=0.1.21
langgraph==0.3.28
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.4.59
llvmlite==0.44.0
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
motor==3.7.1
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nomic==3.4.1
numba==0.61.0
numpy>=1.24.0,<2.2.0
oauthlib==3.2.2
onnxruntime==1.21.0
openai==2.11.0
openai-whisper==20240930
opentelemetry-api==1.32.0
opentelemetry-exporter-otlp-proto-common==1.32.0
opentelemetry-exporter-otlp-proto-grpc==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-proto==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
posthog==3.24.1
primp==0.15.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1>=0.6.1
pyasn1_modules>=0.2.8
pycparser==2.23
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.0
pypdf==5.4.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose>=3.3.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resend==2.10.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9
ruff==0.14.5
safehttpx==0.1.7
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
sib-api-v3-sdk==7.6.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.45.3
sympy==1.13.1
tavily-python==0.5.4
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.15.2
tomlkit==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
transformers==4.39.3
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
cloudinary
